+++ ApplyToBecomeInternal/ApplyToBecome.Data/Models/Application/Loan.cs
@@ -6,10 +6,10 @@ namespace ApplyToBecome.Data.Models.Application
 {
 	public class Loan
 	{
-		public float SchoolLoanAmount { get; set; }
+		public decimal SchoolLoanAmount { get; set; }
 		public string SchoolLoanPurpose { get; set; }
 		public string SchoolLoanProvider { get; set; }
-		public float SchoolLoanInterestRate { get; set; }
+		public decimal SchoolLoanInterestRate { get; set; }
 		public string SchoolLoanSchedule { get; set; }
 	}
 }
